coprocessor-machine/requirements.txt
requests == 2.31.0
pysha3 == 1.0.2
